Revise: further tweaks/fixes to user toolbars
Adds a title to user toolbars with the profile (Host) name so that they are
identifiable if multi-playing with the same toolbar name in each - this
means that they are finally listed in the main window context menu (in a
disabled state as they cannot be hidden from there) instead of the previous
blank, un-checkable, disabled entries.

Add host name to TToolbar objectName so that a packaged one with the same
name in two profiles are distinct - which is a requirement for the
size/position remembering code to work.

Remove pointless and unused:
* (QWidget*) TEasyButtonBar::mpBar
* (QString) TEasyButtonBar::mName

Reduce the number of calls to Host::getName() to a single instance in a few
methods where it is done several times.

Ensure that changing the name of a toolbar in the Editor gets updated in
the toolbar itself.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>