Revise: various small changes prompted during peer review
Change text used when trying to activate an item and it is bugged-out -
(and move the previously bracketed text to a full sentence in a separate
italicised paragraph).

Change the stretch factor used for the Lua script area so it should get
more of the vertical screen space on initial showing.

Fixed a type in one developer's comment in the header texts for the search
results display widget - and change all four to not refer to that as a
"treewidget".

Fixed a spelling type in the color_trigger.ui form!

Restore some tooltips and add another one (for the color highlight option).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>