Revise: add support to EXCLUDE a modifier from a keybinding
The prior code allowed the use of more than one modifier but did not allow
properly for related keybindings to have one with and one without a
particular modifier but with a common another one, e.g. control+somekey
and control+shift+somekey as the first will fire as well when the second is
called by the user.

To do this I have made the six checkboxes on the `keybindings_main_area.ui`
for be tri-stated and revised the code so that they are, by default, in
the partially checked and then treated as "don't cares" like the prior code
did but if unchecked the code checks that they are NOT present in the user
key-press.

The existing Lua functions have been modified to take a further. optional
modifiers argument that specify the modifiers that must NOT be present
and the new `getKeyModifiers()` lua function now returns two values with
the second being the NOT modifiers.  The XML import/export code for
keybindings has been extended to convey this extra data in some previously
unused bits in the Qt::KeyModifiers enum so that older Mudlet version will
not see the NOT modifiers and that newer code that has this feature will
read old saves in a compatible manner -  the data will be lost if saved
from a new version, loaded into an old version, saved from that old version
and then reloaded in the new version, but that is unavoidable...

Should a modifier be specified in both the REQUIRED and NOT REQUIRED field
of a Lua function then the NOT REQUIRED modifier will be silently dropped
when it is stored within the `TKey` instance.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>