Increase cache timeout (#2538)
* Increase cache timeout

Our macOS builds seem to be unable to upload their cache as they run into timeouts. We set the timeout to 10 minutes in the hopes to be able to upload the initial cache in that time and subsequent caches don't need as much time.

* Add the brew cache to the travis cache

See https://stackoverflow.com/questions/39930171/cache-brew-builds-with-travis-ci, this might help speeding up our macOS builds even more.