BugFix: allow packages with sub-directories & get exporter to report result
The previous code would silently fail when using `zip_file_add(...)` if the
item to be added was actually a directory. This commit fixes that by
examining each "file" that is to be added to the archive file and deduces
all sub-directories that need to be made and arranges for those to be
inserted (with `zip_dir_add(...)`) first before populating the archive with
the actual files.

It also refactors the code so that a pass/fail message is displayed at the
end of the package exportation operation - and so that the dialog is not
closed until the user has clicked on a suitable button ("Cancel" before the
package is exported, "Close" if the process failed and "Okay" if it
succeeded). The revised success message also retains a suggestion (with a
clickable link) to upload the ".mpackage" to the appropriate Mudlet forum
topic.

Whilst debugging this I also found:
* `(bool) XMLexport::saveXml(const QString&)` was not returning a `true`
on success - which as the code now checks and reports on failures needed
addressing.
* there were some unused elements on the form/dialog that seemed to be
  left over from a previous implementation that could be removed.
* there was an unused dlgPackageExporter constructor that could be removed
* some class members were not obviously named as if they were - indeed one
  shared the same name as an element on the form/dialog.
* there was an element on the form that could show the target name and path
  of the package but it was being hidden - it is now shown whilst the user
  is selecting the Mudlet items and extra files to include - but is hidden
  on successful package creation as the message for the latter includes the
  same piece of information.
* the dialog that prompts for the package name had a pointless duplication
  of title and prompt texts - the latter has now been revised.
* the dialog that prompts for the package save location was not being
  supplied with a correct argument - instead of a starting path it was
  being given the text "Where do you want to save the package?" which is
  never (with the '?') going to be a valid path!!!
* that same dialog was also not showing the files in the directory - and
  as the code is quite capable of overwriting an existing archive it makes
  sense to show files as well as directories when selection the latter.
  The Qt Documentation suggests that on the Windows platform it is not
  possible to show files in a directory selection dialog using the native
  OS dialogue - I also found this was the case with the KDE Plasma Desktop
  Environment so I have set the options NOT to use a native dialog for all
  platforms, not just the Windows one.

The packages created now have the intended `.mpackage` extension though
they are `.zip` format archives underneath.

This should close https://github.com/Mudlet/Mudlet/issues/2547 and also
close https://github.com/Mudlet/Mudlet/issues/1844 .

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>