Revise: pad RoomId numbers with spaces on 2D map
As requested in the issue below this converts the room Id numbers when
shown in the 2D Mapper to be padded with leading spaces rather than leading
zeroes.

Note that the numbers are shown only when the longest (most digits) Room
Id number for the displayed area can be displayed at a readable size -
otherwise the mapper switches back to displaying room *symbols* if they
will fit inside the room rectangle/circle.

This closes https://github.com/Mudlet/Mudlet/issues/2560 .

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>