BugFix: ensure user windows closure when profile is closed if multi-playing
The previous code in `(void) TConsole::closeEvent(QCloseEvent*)` assumed
the `TConsole` for a user window's object name was what was used as a key
in the profile's main `TConsole`'s `<QString, TConsole*> mSubConsoleMap`
and `<QString, TDockWidget*> mDockWidgetMap`. In fact for user windows the
object name was never set for the `TConsole` and it would not be a good
thing to have both the TDockWidget and the TConsole have the same object
name anyhow.  To further confuse things, the process of constructing the
`TConsole` for a user window did not set a `(QString) mConsoleName` so the
default of "main" was set for that member.

This PR revises the design so that the `TConsole::closeEvent` DOES find
the wanted item in the main `TConsole`'s mSubConsoleMap and mDockWidgetMap
so that the `(TTextEdit*) mUpperPane` and `mLowerPane` DO get the `close()`
method called on them - and there is no error message:
"TConsole::closeEvent(QCloseEvent*) INFO - closing a UserWindow but the
TDockWidget pointer was not found to be removed..." produced for every
user window that is closed when a profile is ended but the Mudlet
application continues to run when multi-playing...

Also:
* A shadowed pair of `TDockWidget` and `TConsole` pointers are no longer
created in `(bool) mudlet::openWindow(...)` by spurious uses of `auto`
(when there are already a pair of `pD` and `pC` in existence).
* Some uses of raw strings or `QString` constructions for `objectName`
are cleaned up to use `QStringLiteral` constructions instead.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>