BugFix: correct some errors
`(bool) Host::stopWatch()` was missing a final `return` value.

The code in `(void) stopWatch::Host::adjustMilliSeconds(const qint64)` used
`(QDateTime) QDateTime::addMSecs(qint64)` incorrectly in that I had thought
it adjusted the QDateTime it was called upon whereas it returns a reference
to the adjusted value - so needs to be invoked in a different manner.

`QString stopWatch::getElapsedDayTimeString() const` made use of `int` for
some intermediate local variables but on Windows platforms the int type may
be 32-bit long and thus not big enough to contain 64-bit values - and some
other locals can be much shorter because of the limits that the code places
on their values but will give compiler warnings without static_cast <T>.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>