Enhance: allow more than one room ID to have same hash value
This PR revises the ability to have a hash (arbitrary QString) to represent
a room number so that the same hash can refer to more than a single room.

The code is refactored in such a way that it should behave in a backwards
compatible manner - i.e. existing packages or scripts should not see any
difference if they invoke the existing functions in the way they have up
to the present. However, if a script or package checks and confirms the
existence of the new function `getRoomIDsbyHash` (note the extra 's') then
it can know that it can invoke the extended functionality of both that
function (which returns a table of ALL the rooms with a hash that matches)
and also:
+ supply an optional `false` third argument to `setRoomIDbyHash` to cause
it to return a nil + error message should the same hash be supplied for
more than one room id - the previous code would silently forget about the
previous room id against that hash and only remember the last one stored.
+ supply an optional `true` second argument to `getRoomIDbyHash` to cause
it to return a nil + error message should there be more (or less) than one
room id stored for the given hash string.  Previous code would return -1
if no room was found and only the last room id stored with that hash if
there was more than one - which is what will happen without that optional
argument being present and true.

Also, `setRoomIDbyHash` can now clear a hash for a room id by supplying an
empty string as the hash. This does mean that no room can have a hash that
is an empty string...

This PR also revises the Mudlet binary Map Format to version 21. This is
necessary and required to retain the `QMultiHash<QString, int>` entity that
now contains the hash to room Id table. This is incompatible with previous
map formats so the map saving code will now warn if a previous format is
used AND there are MULTIPLE ROOMS with the SAME hash. This will not occur
for the prior usage of the hash feature so existing packages/scripts should
not be affected however users that do create hashing schemes (like I have
been doing using WoTMUD's room descriptions) will need to set their map
format control to 21 to prevent data loss (or warnings about it) if they
do have hash collisions.

This PR was prompted by another recent one:
https://github.com/Mudlet/Mudlet/pull/2570 by LiamLeFey
and, as it happens does actually rip out some of his code to do things in
a different way - instead of keeping a central
`QMap<int, QString> TRoomDB::roomIDToHash` to provid the reverse lookup of
room id to hash - the hash value is stored directly within the `TRoom`
instances themselves.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>