Revise: simplify/fix Host::createStopWatch()
The previous code that I wrote to check for a duplicate proposed name for
a stopwatch and also determine the lowest unused positive integer to use
for the id in a single pass was actually defective (and hard to
comprehend after a short time away from it). So I have retreated to some-
thing that is less efficient but is simpler and more readily
understandable.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>