Enhance: allow modification of the player room indication on 2D map
This commit reorganises the 2D Painter code to draw the player room
indicator AFTER all the rooms have been drawn so that it does not get
overdrawn by rooms painted after the player's room - which it could be
if the rooms are at maximum size or the area is in "gridMode".

It also offers some alternative marker options in the form of an oversized
ring around the room which reduces the obstruction of the details of the
room when the marker is drawn on it. The alternatives are:
* a fixed red colour ring
* a fixed blue/yellow colour ring - two contrasting colours mean that it
can not be lost if drawn over a room that already has one of them as its
colour (environment) setting.
* a customizable two colour ring where the user can set the outer and inner
colours as they like - even to the same colour.

This commit does not contain code to save the setting between sessions. It
is intended to get peer/public evaluation of the proposal. A follow up
commit will be appended if this proposal is acceptable and does work.

The ancient `(bool) Host::mMapStrongHighlight` option is still respected,
I note that it was introduced in 3cb1719b0f516eae4fdbd9288e3f8e4e83e9ccbf
from 2011/01/20 (after Mudlet 1.0.5) and the code to control it was lost
in 8b0e8bb0f925d58c558cf864faae6b987435bd82 from 2012/12/31 "NEW Vadim
Peretokin: mapper remembers settings" just before the release of Mudlet
2.0...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>