Add an option to make use of GNU Readline.
GNU readline has a lot of features, so it's nice to use the library
directly rather than try to reimplement it -- inevitably, some
features would be missing.

This patch works with Readline 7.0 on my linux machine.  I don't know
whether it works on platforms (libreadline should be available, in
brew and mingw, but they might be older.)

Known outstanding issues:

- Spellcheck is a bit wonky with punctuation, but that's really an
  existing bug in recheckWholeLine() that can be addressed separately.

- You can't create multiline inputs, and interaction becomes slow for
  very large (>100 word) lines because of spellcheck.

- Readline doesn't behave all that intuitively with a mouse.

- The build only uses -lreadline on unix, not mac or windows.  I have
  no ability to test other platforms.

Only the last one seems serious to me.

Another possible issue is that I reorganized the code a bit even for
when the readline option is disabled.  The only way in which I know
this changes behavior is that now potential user-defined key bindings
take priority over commandLine bindings; in my opinion that is better
behavior, and it made the code simpler.  But it's certainly possible
that I accidentally changed the behavior in other ways -- I don't know
the expected behavior well enough to test it.