Revised: following peer-review hide modifier controls & restore old format
This restored the generation of a string that lists all the modifiers and
the base key - by hiding the modifier controls with a triangular control
that rotates through 90 degrees in the same manner as the one added to
the trigger_main_area form/dialog recently (thought this one is not auto-
mated to response to the splitter movement).

Importantly for the new functionality of being able to require a modifier
to NOT be present to allow the key-binding to fire such a modifier
requirement is also displayed in the text displayed when the modifier
controls are hidden. If a modifier is neither needed or prohibited then
it will not be shown.

The texts used for the modifiers are fully capable of being translated as
there are at least two locales where this will be needed (fr_FR and de_DE)
also the variations between OSes are also taken into account.

There is one key related matter that is not included in this PR which is
the texts for the base keys themselves. Whilst many will not vary between
locales there are some that are not represented by a single letter but are
expressed as an English word which will not work in non-English locales.
However addressing that issue is something that can best be left to a
further PR.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>