Optimization of recheckWholeLine()
recheckWholeLine() spellchecks every word and changes its format.
This appears to take quadratic time.  Unfortunately, there does not
seem to be any way to build a QPlainTextEdit document with different
formatting on every word in a fast way.

This commit makes it so each word's format only needs to be changed if
the word is misspelled -- so it's fast in the common case of having
few typos.