BugFix: get Qts own translations loading
This PR revises the system to load translations to work better. The
code was a rework of some code I prepared some years back, however the
original code was intended to support allowing the changing the UI language
completely dynamically using translation files for Mudlet that could be
located anywhere whereas what was eventually used confined them to be in
a resource file. The restrictions were not completely implemented which
meant that the process to identify what translations were available was
being misused on the Qt system translations and also a code-shortcut
(escaping from test loading of the available Mudlet translations) meant
unused `QTranslator` were needlessly being left in memory.

The details of each language supported by Mudlet have been combined into
a new `translation` class defined in the mudlet .cpp and .h files. This
makes it a simpler process to load the required `.qm` files into
`QTranslator` instances and then those to be installed into the Mudlet
application. (Should the language selection change those then have to be
uninstalled before new ones are put in in their place.)

As a bonus feature the language selection widget in the preferences has
been give the internationally recognise icon developed and licensed under:
Relax-Attribution, Share-Alike, Semi-Noncommercial terms as described
at http://www.languageicon.org/ .

Also, the star rating of translation efforts has been extended to give
a silver-star for translations that are at least three-quarters done
but do not reach the 95% completion for the gold-star.  Unlike the latter
the percentage is still shown.

The revised code is structured so that it could load the translations for
the Mudlet application from somewhere other than the current resources
file (though that is still hard-coded in in this commit) under those
circumstances the Qt lupdate utility that converts the human usable
`.ts` translation files to the binary `.qm` ones that the executable one
needs may not be present and so the translation stats won't be available
either. So, when the translation `.qm` files do not come from the current
resource file a (floppy) disk icon is used for each locale listed in the
control and NO percentage translated figure is shown.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>