Revise: some changes prompted by peer review
Revert addition of silver star for >75% but <95% translations.

Use `std::unique_ptr` for a temporary, heap allocated `QTranslator` used
during scanning for Mudlet and Qt translation files.

Also, connect up a slot in the dlgProfilePreferences class to the new
`(void) mudlet::signal_guiLanguageChanged(const QString&)` signal to ensure
that the "restart Mudlet to complete the language change" message is
presented to the user in the language they have selected rather than the
one in use when the preferences dialogue was opened. It also regenerates
the title of the preferences as the process of the retranslation needed
also reverts the title to the initial state without a profile name shown
even if one was shown before the language/locale was changed.

The two processes - that of calling `retranslateUi(T*)` for a class and
then regenerated texts created programmatically after the class's use
of `setupUi(T*)` pretty much represent the actions that would be needed if
we wanted Mudlet to fully support run-time language/locale changes without
needing a restart at all!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>