Revise: only change Encoder for feedTriggers not whole profile
A recent change to the feedTriggers(...) function changed the behaviour so
that the input data is handled as UTF-8 - whilst this is probably a good
thing for the user it did so by changing all facets of the profile's server
encoding to UTF-8 whilst the feedTriggers(...) function was run and then
put it back to what the user setting was prior to that.

This is possibly not a good idea if data is received from the game server
- or a separate timer sends something to the game server - as both those
events can happen in separate threads (Mudlet is currently set to operate
in a threaded manner) but they still need the encoder settings to be the
specified one.

This PR gets around the problem by using a separate encoder (if necessary)
to convert the supplied text to the required encoding - if a second
optional argument to feedTriggers(...) is set to true.  Additionally this
also prevents the backward compatibility breakage that arose from the
previous change {which was a pain for yours truely who had carefully
constructed some tests that did send non-ASCII data in a specific encoding}
as the original implimentation required and the present Wiki documentation
still states.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>