Enhance: add Discord Ids to some Mudlet Makers in About Dialogue
I suggested this in the Discord "Mudlet-Development" channel and it was
deemed worth trying.  Significantly it uses the same colour for the
Discord link (which seems to have been offered by "Maiyannah" and NOT
Vadim!) as Discord uses (`#7289DA`) as well as for the Discord Ids - this
is so that they can be distinguished from the Github/Gitter Ids and/or
EMail addresses.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>