BugFix: get Qts own translations loading (#2729)
This is a squash and merge of a large number of commits, with the
following (edited) commit messages:

* BugFix: get Qts own translations loading

This PR revises the system to load translations to work better. The
code was a rework of some code I prepared some years back, however the
original code was intended to support allowing the changing the UI language
completely dynamically using translation files for Mudlet that could be
located anywhere whereas what was eventually used confined them to be in
a resource file. The restrictions were not completely implemented which
meant that the process to identify what translations were available was
being misused on the Qt system translations and also a code-shortcut
(escaping from test loading of the available Mudlet translations) meant
unused `QTranslator` were needlessly being left in memory.

The details of each language supported by Mudlet have been combined into
a new `translation` class defined in the mudlet .cpp and .h files. This
makes it a simpler process to load the required `.qm` files into
`QTranslator` instances and then those to be installed into the Mudlet
application. (Should the language selection change those then have to be
uninstalled before new ones are put in in their place.)

As a bonus feature the language selection widget in the preferences has
been give the internationally recognise icon developed and licensed under:
Relax-Attribution, Share-Alike, Semi-Noncommercial terms as described
at http://www.languageicon.org/ .

Also, the star rating of translation efforts has been extended to give
a silver-star for translations that are at least three-quarters done
but do not reach the 95% completion for the gold-star.  Unlike the latter
the percentage is still shown.

The revised code is structured so that it could load the translations for
the Mudlet application from somewhere other than the current resources
file (though that is still hard-coded in in this commit) under those
circumstances the Qt lupdate utility that converts the human usable
`.ts` translation files to the binary `.qm` ones that the executable one
needs may not be present and so the translation stats won't be available
either. So, when the translation `.qm` files do not come from the current
resource file a (floppy) disk icon is used for each locale listed in the
control and NO percentage translated figure is shown.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

Revert addition of silver star for >75% but <95% translations.

Use `std::unique_ptr` for a temporary, heap allocated `QTranslator` used
during scanning for Mudlet and Qt translation files.

Also, connect up a slot in the dlgProfilePreferences class to the new
`(void) mudlet::signal_guiLanguageChanged(const QString&)` signal to ensure
that the "restart Mudlet to complete the language change" message is
presented to the user in the language they have selected rather than the
one in use when the preferences dialogue was opened. It also regenerates
the title of the preferences as the process of the retranslation needed
also reverts the title to the initial state without a profile name shown
even if one was shown before the language/locale was changed.

The two processes - that of calling `retranslateUi(T*)` for a class and
then regenerated texts created programmatically after the class's use
of `setupUi(T*)` pretty much represent the actions that would be needed if
we wanted Mudlet to fully support run-time language/locale changes without
needing a restart at all!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: adjust arguments to QTranslator::load(...)

I was providing a full path and filename as the first argument which does
seem to prevent the tweaking of the filename that Qt uses to derive a
language without a locale type filename to use - importantly this was
breaking the previously working (Linux) loading of the translation files
from a resource file before I started this PR.

This commit contains some debug output lines for comparison with the
installers produced by the CI processes run also on:
https://github.com/Mudlet/Mudlet/pull/2776

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Testing: make Windows CI Build be verbose when using windeployqt

This is so I can attempt to see whether the qt Chinese/Taiwanese
translations are even seen - as they are currently missing from the PR
test builds.

It also temporarily uses an alternate mudlet-installers branch that
has been created to add the same sort of verbosity to the Linux and macOs
installer building - as the Linux ones seem to be missing ALL the Qt
translation files...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revert: remove options for verbose output from installer makers

This stops the temporary use of a branch in the mudlet/installers
repository that was set up to make the `{mac|linux}deployqt` utilities more
verbose...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revert: remove some further debugging messages

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* BugFix: restore some connect(...) lost in git merge from development

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>