Fix incorrect explanation for mMaxRoomIdDigits
The comment despite being 5 lines long does not actually explain what value this field holds! It is also wrong: this field is only ever used once.

I spent 10 minutes double-checking all of the code because of it...

This is why I don't like such long narrative in comments - it forces me to read the comment, it forces me to double-check the code because the comment could be wrong (and in this case, it is), and then I have to spend time reconciling the two and finding out the truth.

I prefer when things are a lot simpler: the comment explains what this field does, and the code explains how this field is used.