Add feature: Ctrl+# tab switching between profiles
This makes is so you can use Ctrl+1 to select the first open profile, Ctrl+2 the second open profile and so on.

This is specicially designed in a way that if the player has already defined Ctrl+1, Ctrl+# as a custom keybinding - that will still work and take priority over this new keybinding; thus this won't break any profiles and annoy players.