Update mudlet.cpp
Adding a call to `pEditor->activateWindow()` so that the window itself isn't just raised, but actually gets activated. Currently, all the buttons on the toolbar to open Triggers/Aliases/Etc are missing this call, meaning that after clicking the button to open the editor, you still have to click the window or your keyboard use goes to the main window instead. I'll have a PR up shortly that addresses the other buttons.