Change "Script Editor" menu entry to respect current view selection (#2987)
* Modifies "Script Editor" entry in the menu bar to keep the current view open, rather than always forcing the view to switch to the Trigger View. This means is the action that the shortcut Alt+E fires.

* Update dlgTriggerEditor.cpp

Fixed coding style issue

* Update mudlet.cpp

Adding a call to `pEditor->activateWindow()` so that the window itself isn't just raised, but actually gets activated. Currently, all the buttons on the toolbar to open Triggers/Aliases/Etc are missing this call, meaning that after clicking the button to open the editor, you still have to click the window or your keyboard use goes to the main window instead. I'll have a PR up shortly that addresses the other buttons.