BugFix: allow for timeStamp region of TConsole to appear as selected
This commit should allow the timestamp region on the left of the display
in a `TConsole` to appear as if it were selected when the first OR last
character in the main text is selected. This is so that selection of one
or more lines of text seem to also include the timestamps - as they will
be if copied and pasted elsewhere.

This might be (at least part of) the solution to:
https://github.com/Mudlet/Mudlet/issues/2978

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>