Revise: provide backwards-compatibility for Lua startStopWatch(id) function
The prior form of startStopWatch(...) would reset and restart the indicated
stopwatch each time that it was called.  This is not really compatible with
the revised functionality which allows the recorded time to be adjusted
even before the stopwatch is first used.  To allow existing scripts to
continue to experience the same API this commit adds an optional second
boolean argument to the `startStopWatch(...)` call ONLY WHEN the first
argument is an id number and NOT a string name.  If the second argument is
omitted (as it will be when using older scripts) or is true then each time
the function is used the stopwatch will be reset and restarted.

If the second argument is false then the new behaviour of starting the
stopwatch only if it is not already running and continuing from whatever
the cumulative past time is on the stopwatch already is carried out - as it
will be if the first argument is a name.

Also remove some extraneous semi-colons I had accidentally put in in the
past...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>