Enhance: add profile's ANSI colours to Lua color_table
This includes ensuring that the first 16 colours that can be redefined by
the user (or, with permission, the Game Server) are also updated when that
happens. This means that UIs can track those colours.

This also revises the handling of the Lua `cecho` function in
`GUIUtils.lua` as that did not previously respond to colour names
containing digits as the `ansi_###` ones included as part of this do.

Also the `color_table` generation process in `GUIUtils.lua` is also
revised so that it does not overwrite an existing `color_table`, instead it
just adds/restores the colours that it previously inserted.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>